Split metric verifier to several files
