+++ src/algorithms/algorithms.h
@@ -17,4 +17,4 @@
 #include "algorithms/association_rules/apriori.h"
 
 /* Metric FD verifier */
-#include "algorithms/metric_verifier.h"
+#include "algorithms/metric/metric_verifier.h"
